+++ packages/typespec-python/src/http.ts
@@ -14,6 +14,7 @@ import {
     SdkServiceResponseHeader,
     UsageFlags,
     SdkHttpOperationExample,
+    SdkHttpErrorResponse,
 } from "@azure-tools/typespec-client-generator-core";
 import {
     camelToSnakeCase,
@@ -101,7 +102,7 @@ function addPagingInformation(
     method: SdkPagingServiceMethod<SdkHttpOperation> | SdkLroPagingServiceMethod<SdkHttpOperation>,
     operationGroupName: string,
 ) {
-    for (const response of method.operation.responses.values()) {
+    for (const response of method.operation.responses) {
         if (response.type) {
             getType(context, response.type)["usage"] = UsageFlags.None;
         }
@@ -164,11 +165,11 @@ function emitHttpOperation(
 ): Record<string, any> {
     const responses: Record<string, any>[] = [];
     const exceptions: Record<string, any>[] = [];
-    for (const [statusCodes, response] of operation.responses) {
-        responses.push(emitHttpResponse(context, statusCodes, response, method)!);
+    for (const response of operation.responses) {
+        responses.push(emitHttpResponse(context, response.statusCodes, response, method)!);
     }
-    for (const [statusCodes, exception] of operation.exceptions) {
-        exceptions.push(emitHttpResponse(context, statusCodes, exception, undefined, true)!);
+    for (const exception of operation.exceptions) {
+        exceptions.push(emitHttpResponse(context, exception.statusCodes, exception, undefined, true)!);
     }
     const result = {
         url: operation.path,
@@ -331,7 +332,7 @@ function emitHttpBodyParameter(
 function emitHttpResponse(
     context: PythonSdkContext<SdkHttpOperation>,
     statusCodes: HttpStatusCodeRange | number | "*",
-    response: SdkHttpResponse,
+    response: SdkHttpResponse | SdkHttpErrorResponse,
     method?: SdkServiceMethod<SdkHttpOperation>,
     isException = false,
 ): Record<string, any> | undefined {
@@ -354,8 +355,8 @@ function emitHttpResponse(
             typeof statusCodes === "object"
                 ? [(statusCodes as HttpStatusCodeRange).start]
                 : statusCodes === "*"
-                ? ["default"]
-                : [statusCodes],
+                  ? ["default"]
+                  : [statusCodes],
         discriminator: "basic",
         type,
         contentTypes: response.contentTypes,